requirements.txt
asgiref==3.7.2
Django==4.2
django-widget-tweaks==1.5.0
sqlparse==0.4.4
tzdata==2023.4
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.7.1
Django==4.2
django-allauth==0.60.0
django-widget-tweaks==1.5.0
idna==3.6
mysqlclient==2.2.1
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.1.0
fontawesomefree==6.5.1